vedhvaani_kundali_api_fixed/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pyswisseph==2.8.0
jinja2==3.1.2
